- silence compiler warnings
git-svn-id: svn://svn.paxle.net/paxle/trunk@2533 f5d3ab42-bacf-4b1a-9692-565be65a0cfe

+++ bundles/Core/src/main/java/org/paxle/core/impl/MWComponent.java
@@ -571,7 +571,7 @@ public class MWComponent<Data> implements IMWComponent<Data>, ManagedService, Me
 			if (deltaDelay != null) {
 				int delay = this.getMaster().getDelay();
 				if (delay < 0) delay = 0;
-				delay += deltaDelay;
+				delay += deltaDelay.intValue();
 				this.master.setDelay(delay);
 			}
 		}
